Upgrade to Bootstrap 4 This commit updgrades our site to the latest version of Bootstrap (version 4).
As part of this upgrade we have carried out a complete audit and rebuild of our site templates, making greater use of _includes to ensure that the various page layouts are consistent, and making it easier to reuse javascript snippets for multiple dashboard pages. The site makes greater use of Bootstrap's responsive breakpoints and responsive utilities, and performance is now significantly better on mobile thanks to the use of table column-hiding and modals. The site now uses a collapsable nav at all display sizes, the content of which is identical on all pages except the data-tracker. Much of the dashboard javascript has been refactored for conistency and performance. This is true for the CSS as well.

On the whole, this commit brings the site to a state of looking pretty dang sharp. There are still a handful of annoyances and enhancments that need to be addressed, all of which will be documented in the respository's issues.